Renamed utility and updated imports
